Add a new task jsExecutionFiles to list *all* JS files sent to JS envs.
This currently includes:

* The resolved JS dependencies (patched for `commonJSName` support)
* The old `scalaJSConfigurationLibs`, which were only used to
  specify one virtual JS file defining the Java system properties
* The linked Scala.js file

Additional JS files can be added (typically at the *front* of the
sequence) in user-space.

This removes one of the last drops of coupling of `jsDependencies`
to the rest of the sbt plugin, which will be needed to extract it
in a separate sbt plugin. Additionally, it simplifies `loadedJSEnv`
so much that there is a good chance we will be able to remove it
completely in the future.