Fix #2800: Refer to the global scope instead of the global object.
This is a major redesign of accesses to "global" variables, both
from a spec point of view and in the implementation.

Previously, Scala.js specified accesses to global variables in
terms of the *global object*. It assumed that it was capable of
accurately detecting the global object of any JS environment, and
worked from there. For example, `js.Dynamic.global.foo` referred
to `$g.foo`, where `$g` was supposed to be the global object.

From a spec point of view, this was therefore different than how
a typical JavaScript code would be written. Indeed, in "normal" JS
we would just write `foo`, and let it be evaluated as a binding in
the *global scope*. This was however fine as, in ECMAScript 5.1,
the global scope and the global object were basically the same
thing.

However, ECMAScript 2015 changed the situation. There are now
bindings in the global scope that are *not* properties of the
global object. Those are `let`, `const` and `class` bindings at
the top-level of scripts. This means that the previous spec of
Scala.js was utterly incapable of referencing those. They were
totally invisible to Scala.js.

Hence, in this commit we completely change the spec of Scala.js,
and consequently the implementation.

Language changes
================

A reference such as `js.Dynamic.global.foo` is therefore specified
as referring to the binding `foo` in the *global scope* of JS. Note
that it is safer than just saying "it is equivalent to writing
`foo` verbatim in JS", as the latter could collide with local
bindings in the scope chain. The spec protects references to the
global scope from any local definition. Objects annotated with
`@JSGlobalScope` are specified similarly to `js.Dynamic.global`.

Although this trivially solves the problem of accessing `let`,
`const` and `class` bindings, this spec change comes with some
backward source incompatible changes.

Compile-time validation
-----------------------

First, only static bindings (known at compile-time) to variables
whose names are valid JS identifiers are allowed. For example, the
following are legal:

    js.Dynamic.global.foo
    js.Dynamic.global.selectDynamic("bar")

but those are illegal:

    js.Dynamic.global.`not-a-JS-identifier`
    js.Dynamic.global.selectDynamic(computedName)

This restriction directly comes from the fact that only statically
known bindings with valid identifier names can be looked up in the
global scope, even in JavaScript.

The particular identifier `arguments` is similarly forbidden, as
it is a magic variable that automatically lexically binds to the
most nested enclosing function. It cannot be looked up in the
global scope in JavaScript either.

Note that `eval` is allowed, but will systematically be protected
as `(0, eval)` when it is *called*, so as to enforce that the
execution of the script is done at the global level, and does not
see (compiler-generated) local bindings.

The second restriction is that references to the global scope,
either as `js.Dynamic.global` or as `GlobalFoo` where `GlobalFoo`
is annotated with `@JSGlobalScope`, cannot be loaded as a *value*.
For example, the following is illegal:

    val x = js.Dynamic.global

This follows from the previous restriction, as references to the
members of such a reference would have to be looked up dynamically,
and that is not possible with the global scope. More generally, it
makes sense since the "global scope" is not a *value* either in
JavaScript, contrary to the global object.

Run-time semantic changes
-------------------------

In addition to the above compile-time restrictions, the spec change
incurs some differences in run-time semantics, when accessing a
global variable that was not declared. For example:

    val x = js.Dynamic.global.doesNotExist

would previously return `undefined`, but now throws a
`ReferenceError`. Similarly, assignments to non-existent global
variables throw a `ReferenceError`, rather than creating the
variable.

This run-time semantic change is the largest source of actual
breakages in user code, as it invalidates the previous idiom used
to perform feature detection in Scala.js. For example:

    if (js.isUndefined(js.Dynamic.global.Promise))

would previously return `true` when `Promise` was not declared,
but will now throw a `ReferenceError`. The fix is to use
`js.typeOf` instead:

    if (js.typeOf(js.Dynamic.global.Promise) == "undefined")

This brings us to the last semantic change, which is that
`js.typeOf` cannot be understood as a *method* anymore, as it does
not follow the evaluation model of Scala, which guarantees that
`method(expr)` is always equivalent to
`{ val x = expr; method(x) }`. With `js.typeOf(expr)`, if `expr`
is a reference to a non-existent global variable, `"undefined"` is
returned, whereas the rewriting throws a `ReferenceError`.

---

Implementation changes
======================

The above changes mostly impact the compiler and the emitter. The
other steps of the pipeline are basically kept intact.

The IR is augmented with an IR node `JSGlobalRef(ident)` which
denotes a reference to a global variable, in the global scope. It
is a valid lhs for the `Assign` node.

The compiler implements the compile-time validation described
above, and emits `JSGlobalRef`s instead of references to
`<linkingInfo>.envInfo.global`. Other than that, nothing changes.

The emitter had to be completely overhauled. The presence of
`JSGlobalRef`s caused significant issues with scoping and naming.

As a naive approach, the translation of `JSGlobalRef(Ident("Foo"))`
is straightforwardly `js.VarRef(js.Ident("Foo"))`. However, this
can cause collisions with two sources of compiler-generated
identifiers:

* Local bindings, such as `VarDef`s and `ParamDef`s.
* Global encoding bindings, such as `$c_sci_List` or
  `$doubleToInt`.

The first group can be dealt with locally in `JSDesugar`. If there
is a reference to a `JSGlobalRef(Ident("Foo"))` in the method, then
a local identifier `Foo` must be renamed. `JSDesugar` proceeds in
two passes to do so, with a mechanism that stops after the first
pass if possible. In the entire test-suite, the second pass is
necessary for only a single method.

The second group is significantly more complicated to address. If
*anywhere in the program*, someone references `$c_sci_List`, then
*everywhere in the program* we need to rename our internal encoding
of `$c_sci_List` into something else (we use `$$c_sci_List`, adding
as many `$`s as necessary to get a non-colliding name).

To do so, `JSDesugar` and `ClassEmitter` keep track of all global
variables that are referenced anywhere, wrapped in `WithGlobals`,
a monadic data structure that eases that propagation. The `Emitter`
invalidates all its caches and re-emits the entire program if the
set of global variables changes in an incremental run, wrt. to the
previous run. In batch-mode, there can therefore be two passes.

To reduce the likelihood of this happening, we eagerly filter out
non "dangerous global refs". A dangerous global ref is one that
starts with a `$` (and is not the single-char string `"$"`).
Indeed, only "dangerous global refs" can collide with our internal
global encodings. In the test-suite, there is not a single
reference to a dangerous global ref, which means that the emitter
can stop after a single pass. We expect most reasonable codebases
to exhibit the same property.

The test for dangerous global refs is therefore put in the
test-suite-ex, so that we do not unnecessarily slow down the
emission of the normal test-suite.