Fix #2953: Avoid calling toString() in primitive isInstanceOf's.
Even though `(v | 0) === v` will indeed return `false` when `v` is
not even a `number`, the part `v | 0` will cause the internal spec
function `ToPrimitive(v, hint Number)` to be called. That function
(section 7.1.1 of the ES 2015 specification) will first try to call
`v.valueOf()` if it exists, and otherwise will try `v.toString()`.

We must avoid this scenario by preemptively ruling out values that
are not `number`s, which we do with a simple `typeof` test.