Fix two quadratic behaviors of fresh name management in the optimizer.
There were two operations related to fresh names that were O(n)
where n is the number of generated fresh names. Since those
operations are called O(m) times, where m is the size of the
generated code, this yielded overall quadratic time.

The first one was fresh name generation itself. The i'th call of
`freshLocalName("foo")` (for a given `foo`) would try `foo`, then
`foo$1`, then `foo$2`, etc. until `foo$i`, which is O(i). We solve
this by remembering  what was the last `i` we had to jump to for
`foo`, so that we can directly start with that one on the next
call. This gives an essentially constant time algorithm.

The second one was during snapshotting of the used names when
starting a new optimistic branch in `tryOrRollback`. To solve this,
we use persistent data structures, i.e., a `var` of type `Map`
instead of a `val` of type `mutable.Map`, so that snapshotting is
constant time.

Together, those fixes bring the pathological case of #2943 from
20 minutes down to 3 minutes.