Fix a quadratic behavior in state backup management in the optimizer.
When starting a new optimistic branch in `tryOrRollback`, we used
to make a backup of all `statesInUse`. This was O(n) in the number
of states in use, giving an overall quadratic behavior.

We fix this using an amortized approach, where making the backups
is O(1), and restoring the backups is amortized O(1). Instead of
maintaining a list of all states in use, and creating backups in
`tryOrRollback`, we make a backup immediately upon chaning the
value of a `State`. Backups are maintained in a list
`stateBackupChain` of all the backups that were made so far. When
we start a new optimistic branch, we take a snapshot of that list
(in O(1)). Upon rollback, we restore all the backups made *since*
that snapshot, which we can compute as the "list diff", i.e., the
prefix of the current `stateBackupChain` that is followed by the
snapshot version. Since we make one backup per state change, and
since each backup is restored at most once, the state changes
pay for making and restoring the corresponding backups, giving an
amortized O(1) implementation.

This brings the pathological case of #2943 from 3 minutes down to
about 5 seconds.