Fix #2921: Forbid SAMs of JS types (except js.{This,}Function).
Since using such a SAM before would produce invalid IR that sent
the optimizer crashing, we directly make this into an *error*,
rather than a warning.

In theory, we could support SAMs of Scala.js-defined JS types,
if they do not have an `apply` method. But that would require more
work in the compiler backend. This commit does not attempt to do
so.