[no-master] Add an option -P:scalajs:sjsDefinedByDefault.
When enabled, this option switches the default for js.Any things
from `@js.native` (0.6.x style) to `@ScalaJSDefined` (1.x style).
The option only applies to source files in the current run, not to
things loaded from the classpath.

This option is intended as a nicer transition between 0.6.x and
1.x. For a time, users will want to cross-compile their codebases.
But there is no way to have Scala.js-defined JS classes that
compile without deprecation warnings under the two versions. This
option fixes that. The codebase can be written in 1.x style, and
the option activated when `scalaJSVersion.startsWith("0.6.")`.

The content of this commit should not reach the `master` branch.