Always assume that source maps will be needed in LinkerFrontend.
The only reason that the frontend wants to know whether source maps
will be emitted or not is to decide whether to take tree positions
into account when detecting changes. This is supposed to somewhat
speed up incremental runs.

Now, it is very rare to disable source maps in fastOpt; and fullOpt
runs, even if incremental in our pipeline, are bottlenecked in
Closure anyway. Therefore, the benefit of this "optimization" is
most likely negligible, and is not worth the complexity.