Remove `jsDependencies` from the bootstrap test.
To do this without implementing ad hoc resolution of `jszip.js`,
we turn the bootstrap test itself into a CommonJS module and
`@JSImport` it.

In turn, this requires that we install `jszip` through npm, which
we automate with a `package.json`. We use the opportunity to also
install the two other more-or-less optional npm packages that we
use: `jsdom` and `source-map-support`.

This finally allows to remove the `JSDependencies` from the
`toolsJS` project, which was the last reference to it in our build.