Add `js.special.globalThis` to access the global `this` value.
In 1.x, because of #2800, Scala.js will not offer any built-in
mechanism to detect the global object of JavaScript. Instead, user
code should be written to do so. That is only possible if we give
access to the global `this` value, as it appears to the top-level
of the .js file produced by Scala.js.

This commit adds a "primitive" `js.special.globalThis` to that
effect. Technically speaking, it is not actually a primitive, as
it is simply implemented as a linking info property.

Introducing `globalThis` in 0.6.x is necessary for codebases that
want to cross-compile between 0.6.x and 1.x, and need to detect the
global object.