Fix #2979: Allow `%%%` in builds that contain no Scala.js project.
In a build that declares a dependency on `sbt-scalajs`, but does
not actually *enable* the `ScalaJSPlugin` in any project, its
`globalSettings` will not be considered. This means that the
setting `isScalaJSProject` is not initialized, and the `%%%` macro
produces references to an undefined setting.

We fix this by declaring a separate `AutoPlugin` that is *always*
triggered, and only initializes that setting. That plugin should
never be referred to in user code, so we put it in an `internal`
package. Additionally, we deprecate it "since forever" so that
users do not refer to it by accident in a `.sbt` file (which will
import it due to it being an `AutoPlugin`).