Fix #2786: Prevent Closure from renaming `$classData`.
This property is used to determine whether an arbitrary JS object
is a Scala object. If GCC renames it to something too common (e.g.,
a single letter like `b`), we can erroneously identify a JS object
as being a Scala object, and then things go wrong.

This slightly increases the size of `-opt.js` files, but it seems
that gziped sizes are virtually not affected.