Deprecate the conversion from `Long` to `js.Any`.
That conversion is not lossless, as it coerces the `Long` into a
`Double` to do so. We deprecate the conversion so that it does not
happen by mistake.

We mark the conversion as deprecated since "forever" rather than
"1.0.0", as a signal that it should never be removed. Indeed, if
the conversion is removed, Scala's weak conversion from `Long` to
`Double` will kick in anyway, also producing the lossy behavior.

We also mark two existing similar conversions in `JSNumberOps` as
deprecated "forever", for the same reason.