Turn most native JS traits in the library into non-native JS traits.
For most of those, it does not make any practical difference, as
they were sealed anyway (or we sealed them in the process). But it
will be easier to convince people that native JS traits should
typically not exist if we do not declare that many ourselves.

Were exempted the traits that define an `apply` method, as those
cannot be non-native.