Recognize any "standard" JVM-style main object in the sbt plugin.
Instead of recognizing objects extending `js.JSApp` as main objects
in the sbt plugin, we now recognize any object with a JVM-style
`main` method (with an `Array[String]` parameter). We use the new
`ModuleInitializer.mainMethodWithArgs` to call them.

This has the nice benefit that we can finally write a
cross-platform main object.

For backward source compatibility, we add in `js.JSApp` a forwarder

    def main(args: Array[String]): Unit = main()

which allows existing `js.JSApp` to still be recognized and run.
`js.JSApp` is nevertheless deprecated, as there is no reason to
use it anymore.