Systematically use the `js.` prefix in the `js` package.
Within the `scala.scalajs.js` package, bare names such as `Any`
refer to `js.Any`. It is however confusing, as we usually expect
such bare names to refer to `scala.Any`.

In this commit, we systematically and consistently write `js.X`
everywhere, except in the very definition of `X`, where it is not
allowed.