Introduce the `WithGlobals` monad around trees in the emitter.
The next commit will need to keep track of a set of global variable
names referred to by a `js.Tree`. When combining trees, the sets
must be union'ed together and propagated. The need for this is
particularly present in the `ClassEmitter`.

To ease this propagation, we introduce the monad `WithGlobals`,
which encapsulates the association of a set of global variables.
However, in this commit, the set is actually always empty. It will
become non-empty in the next commit.

Almost only `ClassEmitter` is affected by the rewrite, but some of
its interfaces are too. `FunctionEmitter` now returns a
`WithGlobals[js.Function]`, and some of `JSGen`'s methods return a
`WithGlobals[js.Tree]`. The `Emitter` is also adapted to receive
`WithGlobals[_]` from the `ClassEmitter`.

This commit cannot be understood on its own; it needs to be put in
the context of the next commit (about referring to the global
scope). This commit isolates the non-functional changes involving
`WithGlobals[_]`.