Introduce the surface of the new linker API.
This commit introduces the portion of the new linker API described
in #2998 that is targeted for the 0.6.x series. It contains:

* `linker.StandardLinker._`
* `linker.Semantics`
* `linker.ModuleKind`
* `linker.standard.OutputMode`
* `linker.standard.StandardLinkerConfigStandardOps`

This is enough to implement the CLI, the sbt plugin, the `QuickLinker` and
partest's `MainGenericRunner` entirely in terms of the new API.

We deprecate `Linker.apply` in favor of `StandardLinker.apply` in order to
nudge other users of the linker towards the new API, assuming they use that
"easy" entry point (instead of manually manipulating front-end and back-end).