Expose `StandardLinker.Config` as an sbt setting.
This commit introduces the new sbt setting `scalaJSLinkerConfig`,
which directly exposes the `StandardLinker.Config` used to create
the `scalaJSLinker`.

This setting replaces the bag of settings that we had before:
* `scalaJSSemantics`
* `scalaJSModuleKind`
* `scalaJSOutputMode`
* `emitSourceMaps`
* `relativeSourceMaps`
* `scalaJSOutputWrapper`
* `scalaJSOptimizerOptions`

For backward compatibility reasons, we must preserve the behavior
of any sbt build that sets the old options, in any arbitrary ways
and in any scopes. At the same time, we want the new sbt settings
to be usable, at least in typical scenarios.

We choose to support setting the new API in two places:
* In the project scope, without any configuration or task axis
* In the fully qualified scope, with all of project, configuration
  and task specified

To achieve this, we derive the old config settings from the new
config settings in the project scope, and derive back the new
config settings from the old config settings in the fully qualified
scope.