Fix #2945: Leave it to sbt to buffer logs
When we created the test adapter, we implemented log buffering ourselves
to prevent interleaved output when running in parallel. However, this is
not necessary since sbt implements its own log buffering. To make
matters worse, this interfered with it.

We fully remove log buffering in the test adapter while keeping an
escape hatch (via Java system properites) in case something goes wrong.