Move FrameworkDetector and HTMLRunnerTemplate to test-adapter.
These two services were `private[sbtplugin]`, but a) they could be
needed by other build tools and b) they require cooperation from
the test-interface. Therefore, it is pretty clear that they must be
public services of the test-adapter.

Moving them required significant amendments, because they relied on
`sbt.TestFramework` and `sbt.TestDefinition`, which are not
available in the context of the test-adapter.