Get rid of of `linkingInfo.envInfo`.
Now that the only places reading `environmentInfo` are in Scala.js
code (and not in the `Emitter` or in `scalajsenv.js`), we can
get rid of it as a "primitive" known by the emitter.

Instead, we access `__ScalaJSEnv` as a global variable from
user-space code in the library.

This means that `__ScalaJSEnv` is still a magic variable from the
standard library point of view, but at least it isn't so from the
*language* point of view anymore.