Enable Scalastyle in shared source directories.
Sadly, the Scalastyle sbt plugin only processes files in the
`scalaSource` directory, rather than in all
`unmanagedSourceDirectories`. This means that our shared sources
were never checked by Scalastyle.

We fix this in this commit. The Scalastyle sbt plugin still being
a non-Auto plugin from sbt < 0.13.5, there are a few complications.
The problem is that settings of such plugins are applied *after*
those defined in `.settings(...)`, hence overriding them (instead
of the other way around). We have to resort to the `.settingSets`
hack that still exists in 0.13. It won't exist in sbt 1.x, but
then neither will non-Auto plugins, so by the time we get there,
Scalastyle will have to use an `AutoPlugin`.