Fuse `testHtml{Fast,Full}Opt` in a unique `testHtml` task.
The new task honors `scalaJSStage` to decide which stage to use.

This gets rid of 2 hacks in our repo: one in our build, and one in
the sbt plugin itself. It also allows `testHtml` to correctly use
`jsExecutionFiles` that are dependent on the `scalaJSStage`, such
as those produced by the `jsDependencies` plugin. Or more
generally, correctly integrate with any custom tasks that produce
different results depending on the stage.

The new behavior also makes more sense when we regard the HTML
test runners as a peculiar form of JS envs. In that sense,
`testHtml` is more similar to `test` than it was before.