Give some discipline to fetching the `.value`s in `{fast,full}OptJS.
We now make it very explicit that all `.value`s performed by
`{fast,full}OptJS` are within the `Def.taskDyn { ... }`, rather
than within the anonymous `Def.task { ... }` inside. This is very
important for `inspect tree` to be meaningful. Previously we were
accessing `(scalaJSLinker in key).value` from the inner task, and
it would not show up in `inspect tree`.