Replace `Def.taskDyn` by `Def.settingDyn` where possible.
Regardless of whether the *result* key is a setting key or a task
key, it is possible to use `Def.settingDyn` as long as the dynamic
part (the part that decides what `Initialize` to return) only
depends on settings. That's because a `TaskKey[T]` is nothing more
than a `SettingKey[Task[T]]`, so the type parameter of `settingDyn`
can be instantiated to `Task[T]` without any problem.

Assuming https://github.com/sbt/sbt/issues/3258 is eventually
implemented, using `Def.settingDyn` will provide a better sbt
experience in `inspect tree`.

Not all `taskDyn`s are removed, but the only remaining ones are
either

* `{fast,full}OptJS`, which do not actually have any dynamic
  dependency, so they already show up nicely in `inspect tree`, or
* things related to the launcher generation, which will not survive
  in Scala.js 1.x.