Fix #3013: Use `genExpr()` for the qualifier of super calls.
Since Scala 2.12, the `qual` part of a super call is not always
`this`. It can be `this.$outer()` or similar things. Therefore, we
cannot blindly assume that the receiver is `this`, nor that the
super call is a JS super call iff the current class is a JS class.

Instead, we correctly use `genExpr(qual)` and inspect
`isRawJSType(qual.tpe)`.