Remove the scalaJSModuleKind sbt task.
It was only used by `loadedTestFrameworks`, so we move its
computation there. Note that, in the process, we "lose" the
dynamism of the computation, which means that even in `NoModule`,
we will force evaluation of `scalaJSLinkedFile`. However that
seems irrelevant as it is also (most likely) evaluated because of
the dependency on `jsExecutionFiles`.