(Don't) fix #3024: Drop support for Scala 2.12.0.
That version suffers from a regression for fields of traits in some
situations, which in particular affects all fields of JS traits
with `= js.undefined` since `js.UndefOr[A]` is an alias for
`A | Unit`.

Since it comes to a choice between supporting that version of
Scala and the ability to define `js.UndefOr[A]` as such an alias,
we favor the latter.