Add an RPC layer to the test interface
This replaces the ad-hoc message passing with a simple RPC layer and is
the first step in implementing #2645. It also mostly removes
dependencies on tools.json by replacing the serialization with a binary
protocol.

The next steps will be to consolidate functionality into a single JS
controller so that we do not need to restart the JS VMs so often.

Note that this change is potentially dangerous as it exercises
functionality in JSEnvs that has never been used before. Notably:

- Calling {`send`|`close`} and `receive` concurrently.
- Relying on blocking calls to `receive` to throw if another thread
  calls `close`.

Since in the 0.6.x branch most of the enviroments are still in the core
repository, the risk is minimal.