Review all the Configs in the linker.
This is a first step towards the new linker API described in #2998.
The classes and interfaces are really touched yet, but all the
`Config` classes have been reviewed, as well as the constructors
using them.

Here are some highlights.

The class `standard.CoreSpec` aggregates the "core specifications"
of a Scala.js linker. At the moment, this contains the `Semantics`,
the `ModuleKind` and the `OutputMode`. These are all the settings
that can affect how the generated `.js` file behaves in an
observable way (for some definition of "observable", e.g., we
consider here that performing better or worse is not observable).

The class `standard.CommonPhaseConfig` contains configuration that
is potentially relevant to any linker phase. This includes future
linking plugins. They currently contain the core specs as well as
the parallel and batch mode flags.

`ESLevel` is gone. We always use `OutputMode` in places where
`ESLevel` was previously used.

`ClearableLinker` does not check consistency over time of the
`GenLinker`s returned by `newLinker()`. This was not possible
anymore since `GenLinker` does not expose any information allowing
it to do so. Anyway, `ClearableLinker` does not have internal
consistency problems even if `newLinker()` would return
inconsistent linkers, so performing the check does not appear to
be necessary.