Get rid of the workaround to find export annotations on `accessed`.
A long time ago, `@JSExport` used not to have
`@field @getter @setter`, and that meant we needed a workaround to
go look in `sym.accessed` for annotations if `sym` was an accessor.
Now that we break binary compatibility, this is not necessary
anymore.

Removing the workaround revealed that we had similarly forgot to
add the meta-annotations to `@JSExportStatic`, which we also fix
here.

The two test that are removed were specifically designed to
simulate a scenario in which we were loading things compiled by an
old version. They are not relevant anymore.