Use a fixed structure rather than functions for runtime class name.
Arbitrary functions cannot be compared for equality, which means
that currently, there is no way to correctly determine whether two
instances of `Semantics` are equivalent. This is an issue for
caches, which cannot invalidate their outputs if the input
`Semantics` have changed.

Moreover, the existing `Semantics.runtimeClassName` field was a
function from `LinkedClass` to `String`, but `LinkedClass`es are
a concept of `linker.standard._` whereas `Semantics` is in
`linker._`.

We replace the unconstrained `LinkedClass => String` function by
an ADT `RuntimeClassNameMapper`. Ultimately, it will only have 3
possibilities, plus an `andThen` combinator:

* Keep all class names unchanged
* Discard all class names and replace everything by `""`
* Regular expression-based replacements

In 0.6.x, however, we have a "custom" combinator to keep supporting
the deprecated arbitrary `LinkedClass => String`.