Fix #3054: Allow rest parameters in `Closure`s in the IR.
It turns out that most of the linker was already perfectly happy
with rest parameters in `Closure`s. In particular, the emitter is
unchanged. It was mostly just the IR checker which was overzealous.

The optimizer did make a silent assumption that there was no rest
parameters in `Closure`s when inlining them, so we make sure that
it does not attempt to inline such `Closure`s.