Fix #3060: Do not check field existence for classes with no instances.
If a class has no instances, the `BaseLinker` will trim it down and
remove all its fields. However, it is still possible that there is
some reachable code that reads or writes those fields (although
such code would inevitably crash with an NPE if actually executed).
The IR checker must not complain in such cases.

We also add a test for a `Select` on a class that is not even
"needed at all". This worked before, but we had no explicit test
for that scenario.