Fix #3055: Correct codegen for super in anonymous JS classes.
In a named JS class, super method calls (resp. super property
accesses) are translated to `JSSuperBracketCall` (resp.
`JSSuperBracketSelect`). These IR nodes receive as first parameter
a reference to enclosing class, so that they can tell in which
prototype they have to start the lookup.

This translation is not valid in anonymous JS classes, because the
enclosing class becomes an `AbstractJSType` rather than a
`JSClass`. Instead, we now eagerly desugar the nodes in the
compiler back-end. This strategy is consistent with how anonymous
JS classes are otherwise translated: they are purely a compiler
back-end concept, and require no support from the IR.

To translate super property accesses (reads and writes), we need
run-time helpers, which we add in `scala.scalajs.runtime`. Their
code is unfortunately duplicated with similar helpers used by the
emitter to translate `JSSuperBracketSelect`s.