Consider `UndefinedParam` as a "transient" IR node.
A transient IR node cannot be serialized nor hashed. This moves the
dangling `UndefinedParam` detection earlier in `GenJSCode`, from
the time of generating Infos to the time of hashing the members of
a `ClassDef`.

This is necessary because we will not generate Infos in the
compiler anymore.

In the future, we should have a generalized notion of transient IR
nodes so that `UndefinedParam` is not singled out like this.