Fix #3078: New Analyzer, bye bye Infos.
This is a big rewrite of all the infrastructure around `Infos`, and
how they are used by the `Analyzer` to compute the reachability
analysis.

Previously, Infos were a "specified" part of the IR, and were
stored in .sjsir files. This was limiting the evolution of the
reachability analysis in binary compatible releases, as explained
in #3078.

With this commit, Infos become almost an implementation detail of
the `Analyzer`. They are also known from the `BaseLinker` and
`Refiner`, who need to provide inputs to the `Analyzer`, but are
otherwise unknown from the rest of the codebase. Obviously, they
are not stored in .sjsir files anymore.

The `Analyzer` will now lazily read the full `ClassDef`s of classes
it needs, and derive Infos from them at link time. In the Refiner,
Infos are also lazily derived from `LinkedClass`es, whereas they
were eagerly recomputed in `GenIncOptimizer` before. However,
whereas before all Infos on the classpath were read and
deserialized, now the Analyzer is smarter: it only reads trees and
computes Infos for classes that are actually touched by the
reachability analysis algorithm. In other words, only classes that
are "needed at all" are deserialized. In fact, with this commit,
the concept of "needed at all" disappears: classes that appear in
the resulting `Analysis` are by construction "needed at all".

In order for the Analyzer to know its entry points, .sjsir files
store a tiny structure `EntryPointsInfo` containing the encoded
name of the class, and a flag indicating whether the class has any
entry point. That structure is all that is deserialized for all
.sjsir files on the classpath.

For incremental runs, the `BaseLinker` and `Refiner` obviously
cache the computed Infos between runs, and avoid recomputing them
if their inputs (`ClassDef`s or `LinkedClass`es, resp.) have not
changed.

Since Infos become an implementation detail of the `Analyzer` and
its "friends" `BaseLinker` and `Refiner`, the `-i`/`--info` option
of `scalajsp` is removed (both in the CLI and in the sbt plugin).
The `InfoPrinter` class is consequently removed as well.

Moreover, since Infos are computed at link time, there is no point
in *checking* them anymore, which means that the `InfoChecker` is
removed.