Redesign primitive numeric types in the IR.
* Add types for `char`, `byte` and `short` in the IR.
* Remove the subtyping relationships between primitive types, for
  example `int <: double`.

Instead of the subtyping relationships, we introduce explicit
widening conversions as `UnaryOp`s. In order to avoid widening
`int`s to `double`s for the purpose of comparing them, we separate
`Num_<` and friends into `Int_<` and `Double_<`. For floats, we do
introduce the conversions in that case.

There is one subtlety for `char` (`CharType`): it is *not* a
subtype of `any`. That's because `char`s must always be boxed in
`j.l.Character` when they are assigned to an `any`.

This fixes #3085, and provides a cleaner solution to the old
issues #2184 and #2780.