Hijack `j.l.Character` and make `char <: Character` in the IR.
This brings `Char` on par with all the other primitive types as far
as the IR is concerned. A primitive `char` (`CharType`) is a
subtype of `j.l.Character`. This removes the complexity in the
compiler and optimizer related to Chars being "special".

Instead, it is now the Emitter's responsibility to deal with Chars.
Indeed, the JS representation of primitive `char`s is *not*, in
fact, a subtype of the JS representation of `j.l.Character`s. The
root cause being that the `toString()` representation is different.
Therefore, now the Emitter auto-boxes primitive `char`s when used
in a place where a non-`char` is expected.

This makes the Emitter a bit more complex, and definitely acutely
aware of `char`'s. However, it allows the IR to be much cleaner,
with no special primitive type that does not behave like the
others.