Fix #2645: Use less JSEnv runners in TestAdapter
Instead of spawning 1 + (2 + P) * F runners per test run, where

  P: level of parallelism
  F: number of testing frameworks

we will only spawn P runners.

The fundamental change that allows us to do so, is that the JS side of
the test adapter is not tied to a purpose anymore: Before, the JVM side
could either start a master or a slave for a given tesing framework (or
a special purpose short-lived piece of code).

We change the JS side ("bridge") to not have any initial state and be
able to do everything as requested via RPC.

Notably, after this change the bridge
- can detect frameworks,
- is not tied to a specific framework
- can create an arbitrary type (master/slave) of runner.
- can hold runners for multiple frameworks at the same time.

As a consequence, multiple frameworks can share the JS bridges and the
JS VM that has been used for test detection can be reused.

Note that as a side-effect, this also fixes #3143 by abolishing all
timeouts.